+++ apps/client/src/app/services/data.service.ts
@@ -72,13 +72,23 @@ export class DataService {
         ACCOUNT: filtersByAccount,
         ASSET_CLASS: filtersByAssetClass,
         ASSET_SUB_CLASS: filtersByAssetSubClass,
+        DATA_SOURCE: [filterByDataSource],
         HOLDING_TYPE: filtersByHoldingType,
         PRESET_ID: filtersByPresetId,
         SEARCH_QUERY: filtersBySearchQuery,
+        SYMBOL: [filterBySymbol],
         TAG: filtersByTag
       } = groupBy(filters, (filter) => {
         return filter.type;
       });
+
+      if (filterByDataSource) {
+        params = params.append('dataSource', filterByDataSource.id);
+      }
+
+      if (filterBySymbol) {
+        params = params.append('symbol', filterBySymbol.id);
+      }
 
       if (filtersByAccount) {
         params = params.append(