+++ apps/api/src/app/redis-cache/redis-cache.service.mock.ts
@@ -1,13 +1,28 @@
+import { Filter } from '@ghostfolio/common/interfaces';
+
 import { Milliseconds } from 'cache-manager';
 
 export const RedisCacheServiceMock = {
+  cache: new Map<string, string>(),
   get: (key: string): Promise<string> => {
-    return Promise.resolve(null);
+    const value = RedisCacheServiceMock.cache.get(key) || null;
+
+    return Promise.resolve(value);
   },
-  getPortfolioSnapshotKey: (userId: string): string => {
-    return `portfolio-snapshot-${userId}`;
+  getPortfolioSnapshotKey: ({
+    filters,
+    userId
+  }: {
+    filters?: Filter[];
+    userId: string;
+  }): string => {
+    const filtersHash = filters?.length;
+
+    return `portfolio-snapshot-${userId}${filtersHash > 0 ? `-${filtersHash}` : ''}`;
   },
   set: (key: string, value: string, ttl?: Milliseconds): Promise<string> => {
+    RedisCacheServiceMock.cache.set(key, value);
+
     return Promise.resolve(value);
   }
 };