+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved the usability of the toggle component
 - Switched to the accounts endpoint in the holding detail dialog
 
 ## 2.107.1 - 2024-09-12